--- src/pages/MyPage/MyPageInfo.tsx
@@ -1,5 +1,0 @@
-function MyPageInfo() {
-  return <div>내 정보 관리</div>;
-}
-
-export default MyPageInfo;
