+++ src/pages/MyPage/MyPageInfo/index.tsx
@@ -1,0 +1,11 @@
+import MyPageSubtitle from '../MyPageSubtitle';
+
+function MyPageInfo() {
+  return (
+    <div>
+      <MyPageSubtitle subtitle="내 정보 관리" />
+    </div>
+  );
+}
+
+export default MyPageInfo;
