--- src/pages/MyPage/MyPageReview.tsx
@@ -1,5 +1,0 @@
-function MyPageReview() {
-  return <div>내 리뷰</div>;
-}
-
-export default MyPageReview;
