+++ src/pages/MyPage/MyPageReview/index.tsx
@@ -1,0 +1,11 @@
+import MyPageSubtitle from '../MyPageSubtitle';
+
+function MyPageReview() {
+  return (
+    <div>
+      <MyPageSubtitle subtitle="내 리뷰" />
+    </div>
+  );
+}
+
+export default MyPageReview;
